requirements.txt
paramiko
pycrypto
click
pyinotify
keyring
keyrings.alt